getting closer
I was wrong. We dont want to get rid of interrupts because thats how we keep track of our encoder changes. What I was trying to say is that we dont need to check if the current value has changed from the previous value. That was just in place to not clutter the serial, but now that we print every specified timeframe that will do the job.